% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{copy_annotation_files_to_extdata}
\alias{copy_annotation_files_to_extdata}
\title{Copy Remote Annotation Files to Local file system}
\usage{
copy_annotation_files_to_extdata()
}
\description{
This function copies a list of remotely available curated annotation files to the
local files system for use with worm_cat_fun.
}
