% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{get_local_annotation_file_names}
\alias{get_local_annotation_file_names}
\title{List Local Annotation Files}
\usage{
get_local_annotation_file_names()
}
\description{
This function returns a list of locally available curated annotation file names for use with worm_cat_fun.
}
\examples{
get_local_annotation_file_names()
}
