% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{get_remote_annotation_file_names}
\alias{get_remote_annotation_file_names}
\title{List Remote Annotation Files}
\usage{
get_remote_annotation_file_names()
}
\description{
This function returns a list of remotely available curated annotation file names for use with worm_cat_fun.
}
\examples{
get_remote_annotation_file_names()
}
