% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{worm_cat_fun}
\alias{worm_cat_fun}
\title{Worm Cat Function}
\usage{
worm_cat_fun(
  file_to_process,
  title = "rgs",
  output_dir = NULL,
  rm_dir = FALSE,
  annotation_file = "whole_genome_v2_nov-11-2021.csv",
  input_type = "Sequence.ID",
  zip_files = TRUE
)
}
\arguments{
\item{file_to_process}{The file name to the gene set to be processed.}

\item{title}{The title for the bubble plots.}

\item{output_dir}{The output directory (this is usually a relative directory from the current working directory).}

\item{rm_dir}{A flag to remove the processing directory, the Default is \code{TRUE}. This assumes the \code{zip_files} is also \code{TRUE} CAUTION: If \code{rm_dir=TRUE} and \code{zip_files=FALSE} no output will be created.}

\item{annotation_file}{An internal annotation file name or a path to an external annotation file.}

\item{input_type}{The type of input data. \code{'Sequence.ID'} or \code{'Wormbase.ID'} the default is \code{'Sequence.ID'}.}

\item{zip_files}{A flag to create a zip file of the final content. The default is \code{TRUE}.}
}
\description{
worm_cat_fun uses a concise list of nested categories defined in the annotation_file
 where each gene is first assigned to a category based on physiological function, and then to a molecular function or cellular location.
 worm_cat_fun output provides a scaled bubble chart that allows the visualization and direct comparison of complex datasets.
 The tool also provides csv files containing input gene annotations, P-values from Fisher’s exact tests, and Bonferroni multiple hypothesis testing corrections.
}
\examples{
worm_cat_fun(file_to_process="/home/rstudio/examples/sams-1_up.csv",output_dir="./output",annotation_file="whole_genome_v2_nov-11-2021.csv",input_type="Wormbase.ID")

}
