% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{get_available_annotation_files}
\alias{get_available_annotation_files}
\title{This function returns a list of available annotation files.}
\usage{
get_available_annotation_files()
}
\description{
This function returns a list of available annotation files.
}
\examples{
get_available_annotation_files()
}
\keyword{wormcat}
