% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm_cat_controller.R
\name{worm_cat_fun}
\alias{worm_cat_fun}
\title{Worm Cat Function}
\usage{
worm_cat_fun(
  file_to_process,
  title = "rgs",
  output_dir = NULL,
  rm_dir = FALSE,
  annotation_file = "whole_genome_v2_nov-11-2021.csv",
  input_type = "Sequence.ID",
  zip_files = TRUE
)
}
\arguments{
\item{file_to_process}{the file name to be processed (file should be in current working directory)}

\item{title}{the title for your bubble plots}

\item{output_dir}{the output directory (is a relative directory from the current working directory)}

\item{rm_dir}{Boolean If FALSE do not remove temp dir. If TRUE remove temp directory}

\item{annotation_file}{provide an internal annotation file name or a path to an external file}

\item{input_type}{'Sequence.ID' or 'Wormbase.ID' the default is Sequence ID}
}
\description{
This function takes a regulated gene set and a category annotations file and runs a Fisher test.
}
\examples{
worm_cat_fun()
}
\keyword{wormcat}
